Move files, minor refactoring
