+++ src/WinRT.Runtime/Interop/EventSource{TDelegate}.cs
@@ -2,11 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading;
 
 namespace WinRT.Interop
 {
@@ -129,45 +125,6 @@ namespace WinRT.Interop
             state = null;
 
             return false;
-        }
-    }
-
-    internal unsafe sealed class EventSource__EventHandler<T> : EventSource<System.EventHandler<T>>
-    {
-        internal EventSource__EventHandler(IObjectReference obj,
-#if NET
-            delegate* unmanaged[Stdcall]<System.IntPtr, System.IntPtr, WinRT.EventRegistrationToken*, int> addHandler,
-#else
-            delegate* unmanaged[Stdcall]<System.IntPtr, System.IntPtr, out WinRT.EventRegistrationToken, int> addHandler,
-#endif
-            delegate* unmanaged[Stdcall]<System.IntPtr, WinRT.EventRegistrationToken, int> removeHandler,
-            int index) : base(obj, addHandler, removeHandler, index)
-        {
-        }
-
-        protected override ObjectReferenceValue CreateMarshaler(System.EventHandler<T> del) =>
-            ABI.System.EventHandler<T>.CreateMarshaler2(del);
-
-        protected override EventSourceState<System.EventHandler<T>> CreateEventState() =>
-            new EventState(_obj.ThisPtr, _index);
-
-        private sealed class EventState : EventSourceState<System.EventHandler<T>>
-        {
-            public EventState(IntPtr obj, int index)
-                : base(obj, index)
-            {
-            }
-
-            protected override System.EventHandler<T> GetEventInvoke()
-            {
-                System.EventHandler<T> handler = (System.Object obj, T e) =>
-                {
-                    var localDel = (System.EventHandler<T>)del;
-                    if (localDel != null)
-                        localDel.Invoke(obj, e);
-                };
-                return handler;
-            }
         }
     }
 }
