+++ src/WinRT.Runtime/Projections/INotifyDataErrorInfo.netstandard2.0.cs
@@ -171,7 +171,7 @@ namespace ABI.System.ComponentModel
             _obj = obj;
 
             _ErrorsChanged =
-                new EventSource__EventHandler<global::System.ComponentModel.DataErrorsChangedEventArgs>(_obj,
+                new EventHandlerEventSource<global::System.ComponentModel.DataErrorsChangedEventArgs>(_obj,
                 (delegate* unmanaged[Stdcall]<IntPtr, IntPtr, out global::WinRT.EventRegistrationToken, int>)(delegate* unmanaged[Stdcall]<IntPtr, IntPtr, global::WinRT.EventRegistrationToken*, int>)_obj.Vftbl.add_ErrorsChanged_1,
                 _obj.Vftbl.remove_ErrorsChanged_2,
                 0);
@@ -211,7 +211,7 @@ namespace ABI.System.ComponentModel
             remove => _ErrorsChanged.Unsubscribe(value);
         }
 
-        private EventSource__EventHandler<global::System.ComponentModel.DataErrorsChangedEventArgs> _ErrorsChanged;
+        private EventHandlerEventSource<global::System.ComponentModel.DataErrorsChangedEventArgs> _ErrorsChanged;
     }
     
     internal static class INotifyDataErrorInfo_Delegates
